# Optimize lifetime tokens