--- a/src/lib.rs
+++ b/src/lib.rs
@@ -1066,6 +1066,10 @@ macro_rules! quote_token {
         $crate::__private::push_ident(&mut $tokens, stringify!($ident));
     };
 
+    ($tokens:ident $lifetime:lifetime) => {
+        $crate::__private::push_lifetime(&mut $tokens, stringify!($lifetime));
+    };
+
     ($tokens:ident $other:tt) => {
         $crate::__private::parse(&mut $tokens, stringify!($other));
     };
@@ -1279,6 +1283,10 @@ macro_rules! quote_token_spanned {
 
     ($tokens:ident $span:ident $ident:ident) => {
         $crate::__private::push_ident_spanned(&mut $tokens, $span, stringify!($ident));
+    };
+
+    ($tokens:ident $span:ident $lifetime:lifetime) => {
+        $crate::__private::push_lifetime_spanned(&mut $tokens, $span, stringify!($lifetime));
     };
 
     ($tokens:ident $span:ident $other:tt) => {